Do not move cursor in python-indent-line
After indenting a line, cursor was moved to the beginning of the
indent (at the first non-space character in the line).  This change
prevent this movement and make the relative position of the cursor
from the indent the same after indenting line.  This behavior is
consistent with other major mode such as emacs-lisp-mode.